eQuilibrator.jl

eQuilibrator.jl is a lightweight interface to the Python package eQuilibrator

Function reference

eQuilibrator.build_rxn_string — Function

build_rxn_string(rxn::Reaction, compoundtype="kegg")

Get a reaction in string format for Equilibrator.

eQuilibrator.map_gibbs_external — Method

map_gibbs_external(fluxres::Dict{String, Float64}, gibbs)

Calculate the Gibbs free energy change taking only the external fluxes into account. NB: you need to account for the biomass function separately.

eQuilibrator.map_gibbs_internal — Function

map_gibbs_internal(fluxres, gibbs, biomassid="BIOMASS")

Calculate the Gibbs free energy change taking only the internal fluxes into account. NB: you need to account for the biomass function separately. NB: the missing fluxes will pick up transporters...

eQuilibrator.map_gibbs_rxns — Method

map_gibbs_rxns(rxns::Array{Reaction, 1}; dgtype="zero", ph=7.0, ionic_str="100 mM", usekegg=true)

Return a dict of rxn.id => ΔG of the specified dgtype. Takes as inputs an array of reactions, rxns, and optional keyword arguments dgtype, which specifies which type of ΔG calculation should be returned. Valid options for dgtype are "phys" and "prime" (anything else is "zero"). Ionic strength can be set through ionic_str which takes a string input, e.g. "150 mM". Only BIGG and KEGG metabolite identifiers are supported, i.e. the reaction needs to have a KEGG or BIGG id listed in the annotation field in the reaction struct. By default KEGG annotations are used to build the reaction strings that are fed to Equilibrator. Note that the first metabolite id is used.