Function reference

eQuilibrator.Equilibrator — Type

struct Equilibrator

A struct used to reference the equilibrator_api object.

eQuilibrator.Equilibrator — Method

Equilibrator(; ph=7.0, pMg = 3.0, ionic_strength = 0.25M, temperature=298.15K)

A constructor for Equilibrator. Keyword arguments ionic_strength and temperature have units. Any appropriate unit may be used, the conversion is handled internally. This is typically slow to load, ~ 30 seconds, since all the thermodynamic data needs to be loaded into memory.

Example

using eQuilibrator
using Unitful

equilibrator = Equilibrator(pH=6.0, pMg = 2.5, ionic_strength=250.0u"mM", temperature=30.0u"°C")

eQuilibrator.ionic_strength — Method

ionic_strength(equilibrator::Equilibrator, x)

Set the ionic strength of equilibrator to x.

Example

ionic_strength(equilibrator, 100u"mM")

eQuilibrator.ionic_strength — Method

ionic_strength(equilibrator::Equilibrator)

Get the ionic strength of equilibrator.

Example

ionic_strength(equilibrator)

eQuilibrator.pH — Method

pH(equilibrator::Equilibrator, x)

Set pH of equilibrator to x.

Example

pH(equilibrator, 3.5)

eQuilibrator.pH — Method

pH(equilibrator::Equilibrator)

Get the pH from equilibrator.

Example

pH(equilibrator)

eQuilibrator.pMg — Method

pMg(equilibrator::Equilibrator, x)

Set pMg of equilibrator to x.

Example

pMg(equilibrator, 1.2)

eQuilibrator.pMg — Method

pMg(equilibrator::Equilibrator)

Get pMg from equilibrator.

Example

pMg(equilibrator)

eQuilibrator.reset — Method

reset(equilibrator:Equilibrator)

Reset equilibrator back to the default starting state with pH = 7.0, pMg = 3.0, temperature = 298 K and ionic_strength = 0.25 M.

eQuilibrator.temperature — Method

temperature(equilibrator::Equilibrator, x)

Set the temperature of equilibrator to x.

Example

temperature(equilibrator, 40u"°C")

eQuilibrator.temperature — Method

temperature(equilibrator::Equilibrator)

Get the temperature of equilibrator.

Example

temperature(equilibrator)

eQuilibrator.dg_prime — Method

dg_prime(equilibrator, rxn_string; concentrations=Dict{String, Any}(), skip_unbalanced=false, balance_warn=true)

Calculate ΔG' of a reaction described by rxn_string using the settings in equilibrator. Optionally, set the concentrations (abundances) of the species involved in the reaction through concentration. If skip_unbalanced is true then return nothing if the reaction is unbalanced. If balance_warn is false then do not emit a warning when the reaction is unbalanced. Note, concentrations that are smaller than 0.001 mM are capped at 0.001 mM (an eQuilibrator/equilibrator_api) restriction.

Example

equilibrator = Equilibrator()

rxn_string = "bigg.metabolite:atp + bigg.metabolite:h2o = bigg.metabolite:adp + bigg.metabolite:pi"
concens = Dict("bigg.metabolite:atp"=>1u"mM", "bigg.metabolite:adp"=>100u"μM", "bigg.metabolite:pi"=>0.003u"M")
dg_prime(equilibrator, rxn_string; concentrations=concens)

eQuilibrator.ln_reversibility_index — Method

ln_reversibility_index(equilibrator, rxn_string; balance_warn=true, skip_unbalanced=false)

Calculate the log of the reversibility index of a reaction described by rxn_string using settings in equilibrator. If skip_unbalanced is true then return nothing if the reaction is unbalanced. If balance_warn is false then do not emit a warning when the reaction is unbalanced.

Example

equilibrator = Equilibrator()

rxn_string = "bigg.metabolite:atp + bigg.metabolite:h2o = bigg.metabolite:adp + bigg.metabolite:pi"
ln_reversibility_index(equilibrator, rxn_string)

eQuilibrator.physiological_dg_prime — Method

physiological_dg_prime(equilibrator, rxn_string; balance_warn=true, skip_unbalanced=false)

Calculate ΔG' at physiological conditions of a reaction described by rxn_string using the settings in equilibrator. If skip_unbalanced is true then return nothing if the reaction is unbalanced. If balance_warn is false then do not emit a warning when the reaction is unbalanced.

Example

equilibrator = Equilibrator()

rxn_string = "bigg.metabolite:atp + bigg.metabolite:h2o = bigg.metabolite:adp + bigg.metabolite:pi"
physiological_dg_prime(equilibrator, rxn_string)

eQuilibrator.standard_dg_prime — Method

standard_dg_prime(equilibrator, rxn_string; balance_warn=true, skip_unbalanced=false)

Calculate ΔG⁰ at standard conditons of a reaction described by rxn_string using the settings in equilibrator. If skip_unbalanced is true then return nothing if the reaction is unbalanced. If balance_warn is false then do not emit a warning when the reaction is unbalanced.

Note, standard conditions means that all concentrations are set at 1 M.

Example

equilibrator = Equilibrator()

rxn_string = "bigg.metabolite:atp + bigg.metabolite:h2o = bigg.metabolite:adp + bigg.metabolite:pi"
standard_dg_prime(equilibrator, rxn_string)

eQuilibrator._lower_bound — Function

_lower_bound(x, ϵ=0.001u"mM")

Caps the lower bound of a concentration to ϵ. Return ϵ if x is less than ϵ, otherwise x. This function is not exported.

eQuilibrator._parse_reaction_part — Method

_parse_reaction_part(str::String, prefix::String)

Helper function for _parse_reaction_string. This function is not exported.

eQuilibrator._parse_reaction_string — Method

_parse_reaction_string(str::String, prefix::String)

Parse a reaction string and insert the prefix in front of each metabolite. This function is not exported.

eQuilibrator._parse_units — Method

_parse_units(x)

Converts units like kilojoule / mole into kJ/mol. This function is not exported.

eQuilibrator.bigg — Method

bigg(str)

Return a reaction string that has "bigg.metabolite:" appended in front of each metabolite. Respects stoichiometric coefficients. "*" in front of metabolites do not matter, but spaces do matter. So 13pgm is different from 13 pgm but 13*pgm is the same as 13 pgm. A string literal version also exists: bigg"..."

See also kegg, metanetx, chebi.

Example

bigg("atp + h2o = adp + pi")
bigg("atp + h2o = adp + 2*pi")
bigg("atp + h2o = adp + 2 pi")

eQuilibrator.chebi — Method

chebi(str)

Return a reaction string that has "CHEBI:" appended in front of each metabolite. Respects stoichiometric coefficients. "*" in front of metabolites do not matter, but spaces do matter. So 13pgm is different from 13 pgm but 13*pgm is the same as 13 pgm. A string literal version also exists: chebi"..."

See also bigg, metanetx, kegg.

Example

chebi(30616 + 33813 = 456216 + 43474")
chebi(30616 + 33813 = 456216 + 2*43474")
chebi(30616 + 33813 = 456216 + 2 43474")

eQuilibrator.kegg — Method

kegg(str)

Return a reaction string that has "kegg:" appended in front of each metabolite. Respects stoichiometric coefficients. "*" in front of metabolites do not matter, but spaces do matter. So 13pgm is different from 13 pgm but 13*pgm is the same as 13 pgm. A string literal version also exists: kegg"..."

See also bigg, metanetx, chebi.

Example

kegg("C00002 + C00001 = C00008 + C00009")
kegg("C00002 + C00001 = C00008 + 2*C00009")
kegg("C00002 + C00001 = C00008 + 2 C00009")

eQuilibrator.metanetx — Method

metanetx(str::String)

Return a reaction string that has "metanetx.chemical:" appended in front of each metabolite. Respects stoichiometric coefficients. "*" in front of metabolites do not matter, but spaces do matter. So 13pgm is different from 13 pgm but 13*pgm is the same as 13 pgm. A string literal version also exists: metanetx"..."

See also bigg, kegg, chebi

Example

metanetx("MNXM3 + MNXM2 = MNXM7 + MNXM9")
metanetx("MNXM3 + MNXM2 = MNXM7 + 2 MNXM9")
metanetx("MNXM3 + MNXM2 = MNXM7 + 2*MNXM9")

eQuilibrator.@bigg_str — Macro

@bigg_str(str)

Return a reaction string that has "bigg.metabolite:" appended in front of each metabolite. Respects stoichiometric coefficients. "*" in front of metabolites do not matter, but spaces do matter. So 13pgm is different from 13 pgm but 13*pgm is the same as 13 pgm.

See also kegg, metanetx, chebi.

Example

bigg"atp + h2o = adp + pi"
bigg"atp + h2o = adp + 2*pi"
bigg"atp + h2o = adp + 2 pi"

eQuilibrator.@chebi_str — Macro

@chebi_str(str)

Return a reaction string that has "chebi:" appended in front of each metabolite. Respects stoichiometric coefficients. "*" in front of metabolites do not matter, but spaces do matter. So 13pgm is different from 13 pgm but 13*pgm is the same as 13 pgm.

See also bigg, metanetx, kegg.

Example

chebi"30616 + 33813 = 456216 + 43474"
chebi"30616 + 33813 = 456216 + 2*43474"
chebi"30616 + 33813 = 456216 + 2 43474"

eQuilibrator.@kegg_str — Macro

@kegg_str(str)

Return a reaction string that has "kegg:" appended in front of each metabolite. Respects stoichiometric coefficients. "*" in front of metabolites do not matter, but spaces do matter. So 13pgm is different from 13 pgm but 13*pgm is the same as 13 pgm.

See also bigg, metanetx, chebi.

Example

kegg"C00002 + C00001 = C00008 + C00009"
kegg"C00002 + C00001 = C00008 + 2*C00009"
kegg"C00002 + C00001 = C00008 + 2 C00009"

eQuilibrator.@metanetx_str — Macro

@metanetx_str(str::String)

Return a reaction string that has "metanetx.chemical:" appended in front of each metabolite. Respects stoichiometric coefficients. "*" in front of metabolites do not matter, but spaces do matter. So 13pgm is different from 13 pgm but 13*pgm is the same as 13 pgm.

See also bigg, kegg, chebi.

Example

metanetx"MNXM3 + MNXM2 = MNXM7 + MNXM9"
metanetx"MNXM3 + MNXM2 = MNXM7 + 2 MNXM9"
metanetx"MNXM3 + MNXM2 = MNXM7 + 2*MNXM9"